requirements.txt
opencv-contrib-python>=4.8.0
numpy>=1.24.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pyyaml>=6.0
jinja2>=3.1.0
python-multipart>=0.0.6
websockets>=12.0
pytest>=8.0.0
pytest-cov>=4.1.0
ruff>=0.2.0
